+++ sketch-compose-koin/src/commonTest/kotlin/com/github/panpf/sketch/compose/koin/common/test/KoinAsyncImageTest.kt
@@ -17,7 +17,7 @@ import com.github.panpf.sketch.test.utils.LifecycleContainer
 import kotlin.test.Test
 
 @OptIn(ExperimentalTestApi::class)
-class SingletonAsyncImageTest {
+class KoinAsyncImageTest {
 
     init {
         Koins.initial()