+++ sketch-core/src/commonMain/kotlin/com/github/panpf/sketch/cache/internal/MemoryCacheRequestInterceptor.kt
@@ -83,10 +83,12 @@ class MemoryCacheRequestInterceptor : RequestInterceptor {
         val memoryCache = requestContext.sketch.memoryCache
         val memoryCacheKey = requestContext.memoryCacheKey
         val cachedValue = memoryCache[memoryCacheKey] ?: return null
+        val imageInfo = cachedValue.getImageInfo() ?: return null
+        val resize = cachedValue.getResize() ?: return null
         return ImageData(
             image = cachedValue.image,
-            imageInfo = cachedValue.getImageInfo()!!,
-            resize = cachedValue.getResize()!!,
+            imageInfo = imageInfo,
+            resize = resize,
             transformeds = cachedValue.getTransformeds(),
             extras = cachedValue.getExtras(),
             dataFrom = DataFrom.MEMORY_CACHE,
@@ -99,9 +101,23 @@ class MemoryCacheRequestInterceptor : RequestInterceptor {
         val requestContext = chain.requestContext
         val target = request.target
         if (target != null) {
-            val placeholderDrawable = request.placeholder
-                ?.getImage(sketch, request, null)
-                ?.resizeOnDraw(request, requestContext.size)
+            val placeholderDrawable = runCatching {
+                request.placeholder
+                    ?.getImage(sketch, request, null)
+                    ?.resizeOnDraw(request, requestContext.size)
+            }.apply {
+                if (isFailure) {
+                    val exception = exceptionOrNull()
+                    if (exception != null) {
+                        sketch.logger.e(
+                            tr = exception,
+                            msg = "Failed to get the placeholder image. '${request.key}'"
+                        )
+                    } else {
+                        sketch.logger.e(msg = "Failed to get the placeholder image. '${request.key}'")
+                    }
+                }
+            }.getOrNull()
             target.onStart(sketch, request, placeholderDrawable)
         }
     }