+++ sketch-view-koin/src/androidTest/kotlin/com/github/panpf/sketch/view/koin/test/ImageViewKoinExtensionsTest.kt
@@ -41,7 +41,7 @@ import kotlin.test.assertNotNull
 import kotlin.test.assertNull
 
 @RunWith(AndroidJUnit4::class)
-class ImageViewSingletonExtensionsTest {
+class ImageViewKoinExtensionsTest {
 
     init {
         Koins.initial()